Made the wins and losses fields of Player "int"s

+++ src/main/java/org/devnull/matchmaking/PlayerInterface.java
@@ -4,8 +4,8 @@ public interface PlayerInterface {
 
     String getName();
 
-    long getWins();
+    int getWins();
 
-    long getLosses();
+    int getLosses();
 
 }
